New implementation of screens
Updated the implementation of existing screens to use ScreenManager to facilitate screen transitions. Added versions of log in and create account screens for dock workers with a screen to ask if you are a worker after clicking log in or create account from welcome screen. Made an alternate version of link boat screen for use during account creation specifically. Finally, added home screen after logging in with options to schedule a time, link a new boat, contact the dock, and log out, with a screen after clicking log out to confirm. Still missing: most worker-side functions, scheduling time, and contacting dock, as well as linking to database.