changes to createUser function
added parameters to createUser in dbAPI to now take the text from the confirm password field as well as if the user is a worker or not. first checks if there is an existing user with entered username or email, returning false if so. then checks if password and confirm password fields match, returning false if they do not. then checks if the user is a worker, creating a new user with is_worker set to 1 if so and set to 0 if not. createAccount in Main.py then uses return value and isWorker parameter to set the next screen for the application to go to. main.kv now calls createAccount with the text from confirm password and a boolean corresponding to if they are a worker passed in